Update README.md
Space Invaders is a 1978 shoot 'em up arcade game developed by Tomohiro Nishikado. Within the shooter game genre, Space Invaders was the first fixed shooter and set the template for the shoot 'em up genre. The goal is to defeat wave after wave of descending aliens with a horizontally moving laser to earn as many points as possible.